environments: support AMS organizations in OPA-AMS
We agreed with the AMS team on a method for mapping Observatorium
tenants to organizations IDs in AMS, however oganization IDs are not
available anywhere in the JWTs issued by sso.redhat.com. As a result, we
need to maintain an explicit mapping of tenant names to oganization IDs.
This commit adds the Jsonnet code to support dynamically adding mapping
flags to opa-ams based on an object in the configuration.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>