Remove ThanosQueryOverloadAlert
These are being removed for two reasons.

First, given the topology of Querier we have decided to run in our
clusers it will be firing all the time, as we decided to configure them
with a query concurrency of 1.

Second, this isn't an actionable alert. Our Queriers being overloaded
only require an alert in case this is quickly burning our SLO error
budget. For this we have a separate alert.

Signed-off-by: Douglas Camata <159076+douglascamata@users.noreply.github.com>