Filter datasources on Thanos mixin dashboards
With this commit we filter the available datasources on the grafana
dashboards provided by the bundled thanos mixin dashboards. This should
make it significantly easier to find the data needed.

The regex used currently will show the following datasources:
- app-sre-stage-01-prometheus
- rhobs-testing-cluster-prometheus
- rhobs-testing-prometheus
- rhobsp02ue1-cluster-prometheus
- rhobsp02ue1-prometheus
- telemeter-prod-01-prometheus

Any additional datasources starting with `rhobs` will be added if they
become available.

Signed-off-by: Jacob Baungard Hansen <jacobbaungard@redhat.com>